+++ Engine/Source/Runtime/Engine/Private/VT/VirtualTextureLevelRedirector.h
@@ -1,14 +1,11 @@
-// Copyright Epic Games, Inc. All Rights Reserved.
+// Copyright 1998-2019 Epic Games, Inc. All Rights Reserved.
 
 #pragma once
 
 #include "CoreMinimal.h"
 #include "VirtualTexturing.h"
 
-/** 
- * IVirtualTexture implementation that redirects requests to one of two children depending on vLevel.
- * We take ownership of the referenced IVirtualTexture objects meaning that they are destroyed when this object is destroyed.
- */
+/** IVirtualTexture implementation that redirects requests to one of two children depending on vLevel. */
 class FVirtualTextureLevelRedirector : public IVirtualTexture
 {
 public:
@@ -20,7 +17,7 @@ public:
 		const FVirtualTextureProducerHandle& ProducerHandle,
 		uint8 LayerMask,
 		uint8 vLevel,
-		uint64 vAddress,
+		uint32 vAddress,
 		EVTRequestPagePriority Priority
 	) override;
 
@@ -31,7 +28,7 @@ public:
 		const FVirtualTextureProducerHandle& ProducerHandle,
 		uint8 LayerMask,
 		uint8 vLevel,
-		uint64 vAddress,
+		uint32 vAddress,
 		uint64 RequestHandle,
 		const FVTProduceTargetLayer* TargetLayers
 	) override;
